Login and NotFound pages created
The Login component is responsible for handling user authentication, allows users to log into their accounts. It manages state, handles forms and interacts with API to authenticate the user, then navigates user to home page

The NotFound component is a simple component that handles the 404 Not Found page. It informs users that the page they are looking for cannot be found and provides a link to navigate back to the home page.